fix(snap): Update dependencies and versioning
This commit fixes two build issues with the snap.

1) Source tags for iotech-c-utils and device-sdk-c updated
to the latest version. Also update the source tag for mraa
to use the long git commit hash instead of the short hash.

2) A VERSION file is required by the cmake build so it is
created if it doesn't exist.

Snap versioning was also updated to use "snapcraftctl set-version"
and to not use the git tag or date in the version string.

Signed-off-by: Siggi Skulason <siggi.skulason@canonical.com>